Reorganize MongoDB command registration
