+++ src/mongo/commands/createMongoScrapbook.ts
@@ -1,0 +1,10 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import * as vscodeUtil from "../../utils/vscodeUtils";
+
+export async function createMongoSrapbook() {
+    await vscodeUtil.showNewFile('', 'Scrapbook', '.mongo');
+}
