+++ src/mongo/commands/executeAllMongoCommand.ts
@@ -1,0 +1,13 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { IActionContext } from "@microsoft/vscode-azext-utils";
+import { executeAllCommandsFromActiveEditor } from "../MongoScrapbook";
+import { loadPersistedMongoDB } from "../registerMongoCommands";
+
+export async function executeAllMongoCommand(context: IActionContext) {
+    await loadPersistedMongoDB();
+    await executeAllCommandsFromActiveEditor(context);
+}
