+++ src/mongo/commands/launchMongoShell.ts
@@ -1,0 +1,12 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import * as vscode from "vscode";
+
+export function launchMongoShell(): void {
+    const terminal: vscode.Terminal = vscode.window.createTerminal('Mongo Shell');
+    terminal.sendText(`mongo`);
+    terminal.show();
+}
