+++ src/mongo/registerMongoCommands.ts
@@ -3,25 +3,29 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { AzExtTreeItem, callWithTelemetryAndErrorHandling, IActionContext, IErrorHandlerContext, ITreeItemPickerContext, registerCommandWithTreeNodeUnwrapping, registerErrorHandler, registerEvent } from "@microsoft/vscode-azext-utils";
+import { callWithTelemetryAndErrorHandling, IActionContext, IErrorHandlerContext, registerCommandWithTreeNodeUnwrapping, registerErrorHandler, registerEvent } from "@microsoft/vscode-azext-utils";
 import * as vscode from 'vscode';
-import { Experience, MongoExperience } from '../AzureDBExperiences';
-import { cosmosMongoFilter } from "../constants";
 import { ext } from "../extensionVariables";
-import * as vscodeUtil from '../utils/vscodeUtils';
+import { connectedMongoKey, connectMongoDatabase } from "./commands/connectMongoDatabase";
+import { createMongoCollection } from "./commands/createMongoCollection";
+import { createMongoDatabase } from "./commands/createMongoDatabase";
+import { createMongoDocument } from "./commands/createMongoDocument";
+import { createMongoSrapbook } from "./commands/createMongoScrapbook";
+import { deleteMongoCollection } from "./commands/deleteMongoCollection";
+import { deleteMongoDB } from "./commands/deleteMongoDatabase";
+import { deleteMongoDocument } from "./commands/deleteMongoDocument";
+import { executeAllMongoCommand } from "./commands/executeAllMongoCommand";
+import { executeMongoCommand } from "./commands/executeMongoCommand";
+import { launchMongoShell } from "./commands/launchMongoShell";
+import { openMongoCollection } from "./commands/openMongoCollection";
 import { MongoConnectError } from './connectToMongoClient';
 import { MongoDBLanguageClient } from "./languageClient";
-import { executeAllCommandsFromActiveEditor, executeCommandFromActiveEditor, getAllErrorsFromTextDocument } from "./MongoScrapbook";
+import { getAllErrorsFromTextDocument } from "./MongoScrapbook";
 import { MongoCodeLensProvider } from "./services/MongoCodeLensProvider";
-import { setConnectedNode } from "./setConnectedNode";
-import { MongoAccountTreeItem } from "./tree/MongoAccountTreeItem";
-import { MongoCollectionTreeItem } from "./tree/MongoCollectionTreeItem";
-import { MongoDatabaseTreeItem } from "./tree/MongoDatabaseTreeItem";
-import { MongoDocumentTreeItem } from "./tree/MongoDocumentTreeItem";
 
-const connectedMongoKey: string = 'ms-azuretools.vscode-cosmosdb.connectedDB';
 let diagnosticsCollection: vscode.DiagnosticCollection;
 const mongoLanguageId: string = 'mongo';
+let loadPersistedMongoDBPromise: Promise<void> | undefined;
 
 export function registerMongoCommands(): void {
     ext.mongoLanguageClient = new MongoDBLanguageClient();
@@ -33,95 +37,68 @@ export function registerMongoCommands(): void {
     ext.context.subscriptions.push(diagnosticsCollection);
 
     setUpErrorReporting();
+    loadPersistedMongoDB();
 
-    const loadPersistedMongoDBTask: Promise<void> = loadPersistedMongoDB();
+    registerCommandWithTreeNodeUnwrapping('cosmosDB.launchMongoShell', launchMongoShell);
+    registerCommandWithTreeNodeUnwrapping('cosmosDB.newMongoScrapbook', createMongoSrapbook);
+    registerCommandWithTreeNodeUnwrapping('cosmosDB.executeMongoCommand', executeMongoCommand);
+    registerCommandWithTreeNodeUnwrapping('cosmosDB.executeAllMongoCommands', executeAllMongoCommand);
+
+    // #region Account command
 
     registerCommandWithTreeNodeUnwrapping('cosmosDB.createMongoDatabase', createMongoDatabase);
+
+    // #endregion
+
+    // #region Database command
+
+    registerCommandWithTreeNodeUnwrapping('cosmosDB.connectMongoDB', connectMongoDatabase);
     registerCommandWithTreeNodeUnwrapping('cosmosDB.createMongoCollection', createMongoCollection);
-    registerCommandWithTreeNodeUnwrapping('cosmosDB.createMongoDocument', async (context: IActionContext, node?: MongoCollectionTreeItem) => {
-        if (!node) {
-            node = await pickMongo<MongoCollectionTreeItem>(context, MongoCollectionTreeItem.contextValue);
-        }
-        const documentNode = await node.createChild(context);
-        await vscode.commands.executeCommand("cosmosDB.openDocument", documentNode);
-    });
-    registerCommandWithTreeNodeUnwrapping('cosmosDB.connectMongoDB', async (context: IActionContext, node?: MongoDatabaseTreeItem) => {
-        if (!node) {
-            // Include defaultExperience in the context to prevent https://github.com/microsoft/vscode-cosmosdb/issues/1517
-            const experienceContext: ITreeItemPickerContext & { defaultExperience?: Experience } = { ...context, defaultExperience: MongoExperience };
-            node = await pickMongo<MongoDatabaseTreeItem>(experienceContext, MongoDatabaseTreeItem.contextValue);
-        }
-
-        const oldNodeId: string | undefined = ext.connectedMongoDB && ext.connectedMongoDB.fullId;
-        await ext.mongoLanguageClient.connect(node.connectionString, node.databaseName);
-        void ext.context.globalState.update(connectedMongoKey, node.fullId);
-        setConnectedNode(node);
-        await node.refresh(context);
-
-        if (oldNodeId) {
-            // We have to use findTreeItem to get the instance of the old node that's being displayed in the ext.rgApi.appResourceTree. Our specific instance might have been out-of-date
-            const oldNode: AzExtTreeItem | undefined = await ext.rgApi.appResourceTree.findTreeItem(oldNodeId, context);
-            if (oldNode) {
-                await oldNode.refresh(context);
-            }
-        }
-    });
     registerCommandWithTreeNodeUnwrapping('cosmosDB.deleteMongoDB', deleteMongoDB);
+
+    // #endregion
+
+    // #region Collection command
+
+    registerCommandWithTreeNodeUnwrapping('cosmosDB.openCollection', openMongoCollection);
+    registerCommandWithTreeNodeUnwrapping('cosmosDB.createMongoDocument', createMongoDocument);
     registerCommandWithTreeNodeUnwrapping('cosmosDB.deleteMongoCollection', deleteMongoCollection);
-    registerCommandWithTreeNodeUnwrapping('cosmosDB.deleteMongoDocument', async (context: IActionContext, node?: MongoDocumentTreeItem) => {
-        const suppressCreateContext: ITreeItemPickerContext = context;
-        suppressCreateContext.suppressCreatePick = true;
-        if (!node) {
-            node = await pickMongo<MongoDocumentTreeItem>(context, MongoDocumentTreeItem.contextValue);
-        }
-        await node.deleteTreeItem(context);
-    });
-    registerCommandWithTreeNodeUnwrapping('cosmosDB.openCollection', async (context: IActionContext, node?: MongoCollectionTreeItem) => {
-        if (!node) {
-            node = await pickMongo<MongoCollectionTreeItem>(context, MongoCollectionTreeItem.contextValue);
-        }
-        await ext.fileSystem.showTextDocument(node);
-    });
-    registerCommandWithTreeNodeUnwrapping('cosmosDB.launchMongoShell', launchMongoShell);
-    registerCommandWithTreeNodeUnwrapping('cosmosDB.newMongoScrapbook', async () => await vscodeUtil.showNewFile('', 'Scrapbook', '.mongo'));
-    registerCommandWithTreeNodeUnwrapping('cosmosDB.executeMongoCommand', async (context: IActionContext, position?: vscode.Position) => {
-        await loadPersistedMongoDBTask;
-        await executeCommandFromActiveEditor(context, position);
-    });
-    registerCommandWithTreeNodeUnwrapping('cosmosDB.executeAllMongoCommands', async (context: IActionContext) => {
-        await loadPersistedMongoDBTask;
-        await executeAllCommandsFromActiveEditor(context);
-    });
+
+    // #endregion
+
+    // #region Document command
+
+    registerCommandWithTreeNodeUnwrapping('cosmosDB.deleteMongoDocument', deleteMongoDocument);
+
+    // #endregion
 }
 
-async function loadPersistedMongoDB(): Promise<void> {
-    // NOTE: We want to make sure this function never throws or returns a rejected promise because it gets awaited multiple times
-    await callWithTelemetryAndErrorHandling('cosmosDB.loadPersistedMongoDB', async (context: IActionContext) => {
-        context.errorHandling.suppressDisplay = true;
-        context.telemetry.properties.isActivationEvent = 'true';
+export async function loadPersistedMongoDB(): Promise<void> {
+    if (loadPersistedMongoDBPromise) {
+        return loadPersistedMongoDBPromise;
+    } else {
+        // NOTE: We want to make sure this function never throws or returns a rejected promise because it gets awaited multiple times
+        loadPersistedMongoDBPromise = callWithTelemetryAndErrorHandling('cosmosDB.loadPersistedMongoDB', async (context: IActionContext) => {
+            context.errorHandling.suppressDisplay = true;
+            context.telemetry.properties.isActivationEvent = 'true';
 
-        try {
-            const persistedNodeId: string | undefined = ext.context.globalState.get(connectedMongoKey);
-            if (persistedNodeId) {
-                const persistedNode = await ext.rgApi.appResourceTree.findTreeItem(persistedNodeId, context);
-                if (persistedNode) {
-                    await ext.mongoLanguageClient.client.onReady();
-                    await vscode.commands.executeCommand('cosmosDB.connectMongoDB', persistedNode);
+            try {
+                const persistedNodeId: string | undefined = ext.context.globalState.get(connectedMongoKey);
+                if (persistedNodeId) {
+                    const persistedNode = await ext.rgApi.appResourceTree.findTreeItem(persistedNodeId, context);
+                    if (persistedNode) {
+                        await ext.mongoLanguageClient.client.onReady();
+                        await vscode.commands.executeCommand('cosmosDB.connectMongoDB', persistedNode);
+                    }
+                }
+            } finally {
+                // Get code lens provider out of initializing state if there's no connected DB
+                if (!ext.connectedMongoDB) {
+                    ext.mongoCodeLensProvider.setConnectedDatabase(undefined);
                 }
             }
-        } finally {
-            // Get code lens provider out of initializing state if there's no connected DB
-            if (!ext.connectedMongoDB) {
-                ext.mongoCodeLensProvider.setConnectedDatabase(undefined);
-            }
-        }
-    });
-}
-
-function launchMongoShell(): void {
-    const terminal: vscode.Terminal = vscode.window.createTerminal('Mongo Shell');
-    terminal.sendText(`mongo`);
-    terminal.show();
+        });
+    }
 }
 
 function setUpErrorReporting(): void {
@@ -169,55 +146,4 @@ function updateErrorsInScrapbook(context: IActionContext, document: vscode.TextD
     } else {
         context.telemetry.suppressIfSuccessful = true;
     }
-}
-
-export async function createMongoDatabase(context: IActionContext, node?: MongoAccountTreeItem): Promise<void> {
-    if (!node) {
-        node = await pickMongo<MongoAccountTreeItem>(context);
-    }
-    const databaseNode = <MongoDatabaseTreeItem>await node.createChild(context);
-    await databaseNode.createChild(context);
-
-    await vscode.commands.executeCommand('cosmosDB.connectMongoDB', databaseNode);
-}
-
-export async function createMongoCollection(context: IActionContext, node?: MongoDatabaseTreeItem): Promise<void> {
-    if (!node) {
-        node = await pickMongo<MongoDatabaseTreeItem>(context, MongoDatabaseTreeItem.contextValue);
-    }
-    const collectionNode = await node.createChild(context);
-    await vscode.commands.executeCommand('cosmosDB.connectMongoDB', collectionNode.parent);
-}
-
-export async function deleteMongoDB(context: IActionContext, node?: MongoDatabaseTreeItem): Promise<void> {
-    const suppressCreateContext: ITreeItemPickerContext = context;
-    suppressCreateContext.suppressCreatePick = true;
-    if (!node) {
-        node = await pickMongo<MongoDatabaseTreeItem>(context, MongoDatabaseTreeItem.contextValue);
-    }
-    await node.deleteTreeItem(context);
-    if (ext.connectedMongoDB && ext.connectedMongoDB.fullId === node.fullId) {
-        setConnectedNode(undefined);
-        void ext.context.globalState.update(connectedMongoKey, undefined);
-        // Temporary workaround for https://github.com/microsoft/vscode-cosmosdb/issues/1754
-        void ext.mongoLanguageClient.disconnect();
-    }
-}
-
-export async function deleteMongoCollection(context: IActionContext, node?: MongoCollectionTreeItem): Promise<void> {
-    const suppressCreateContext: ITreeItemPickerContext = context;
-    suppressCreateContext.suppressCreatePick = true;
-    if (!node) {
-        node = await pickMongo<MongoCollectionTreeItem>(context, MongoCollectionTreeItem.contextValue);
-    }
-    await node.deleteTreeItem(context);
-}
-
-async function pickMongo<T extends AzExtTreeItem>(context: IActionContext, expectedContextValue?: string | RegExp | (string | RegExp)[]): Promise<T> {
-    return await ext.rgApi.pickAppResource<T>(context, {
-        filter: [
-            cosmosMongoFilter
-        ],
-        expectedChildContextValue: expectedContextValue
-    });
 }
